Style TractDemo component
