--- client/src/components/Demographics/Demographics.module.scss
@@ -1,18 +1,0 @@
-@use '../../styles/design-system.scss' as *;
-
-.demographicsContainer{
-    padding: 1.2rem 1rem 0 1.2rem;
-
-    .demographicItem {
-        display: flex;
-        justify-content: space-between;
-
-        span {
-            font-size: .8rem;
-            margin-top: 0;
-        }
-    }
-    .customDemographicHeading {
-        @include u-margin(1.5);
-    }
-}
