--- client/src/components/Demographics/index.ts
@@ -1,2 +1,0 @@
-import Demographics from './Demographics';
-export default Demographics;
